Remove unused wallet traits (#1574)
* fix: remove unused wallet traits

* refactor: move Acquired to the Mint where it is used
