+++ crates/cdk/src/mint/proofs.rs
@@ -1,4 +1,5 @@
-use cdk_common::database::{Acquired, DynMintTransaction};
+use cdk_common::database::mint::Acquired;
+use cdk_common::database::DynMintTransaction;
 use cdk_common::mint::ProofsWithState;
 use cdk_common::state::{self, check_state_transition};
 use cdk_common::{Error, State};
